tts/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
piper-tts==1.2.0
numpy==2.1.3
requests==2.32.3
